requirements.txt
bs4
openai
requests
streamlit
python-dotenv